feature: folders decomposition, add @/variables
