+++ app/src/components/MenuGridItem.vue
@@ -17,7 +17,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../variables';
+@import '@/variables';
 
 .menu-list-item {
   display: flex;
