+++ app/src/components/MenuListItem.vue
@@ -18,7 +18,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../variables';
+@import '@/variables';
 
 .menu-item {
   color: $primary-text-color;
