+++ app/src/components/TheFooter.vue
@@ -23,7 +23,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../variables';
+@import '@/variables';
 
 .footer {
   position: absolute;
