+++ app/src/components/cards/TestCard.vue
@@ -55,7 +55,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../variables';
+@import '@/variables';
 
 .test-card {
   display: flex;
