+++ app/src/components/cards/TrainingCard.vue
@@ -41,7 +41,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../variables';
+@import '@/variables';
 
 .training-card {
   position: relative;
